+++ PodoFrontGonnet/src/components/Cards.jsx
@@ -1,12 +1,12 @@
 import Card from "./Card";
-import "../pages/css/cards.css";
+import "./css/cards.css";
 const Cards = ({ listaServicios }) => {
   return (
     <div className="card-section ">
       {listaServicios &&
-        listaServicios.map((e) => e.estado == true ? <Card listaServicios={e} key={e.id} /> : null)}
-
-
+        listaServicios.map((e) =>
+          e.estado == true ? <Card listaServicios={e} key={e.id} /> : null
+        )}
     </div>
   );
 };
