+++ PodoFrontGonnet/src/components/Footer.jsx
@@ -1,5 +1,5 @@
-import React from "react";
-import "../pages/css/footer.css";
+import "./css/footer.css";
+
 const Footer = () => {
   return (
     <footer className="footer">
