+++ PodoFrontGonnet/src/components/ImagenPricipaLogo.jsx
@@ -1,8 +1,12 @@
-import '../pages/css/imagenPricipalInicio.css'
-const ImagenPricipaLogo = () => {
-    return (
-        <img  className="imagenPricipalInicio" src='/src/assets/ImagenesOptimizadas/ImagenPrincipalInicio.png' />
-    )
-}
+import "./css/imagenPricipalInicio.css";
 
-export default ImagenPricipaLogo
+const ImagenPricipaLogo = () => {
+  return (
+    <img
+      className="imagenPricipalInicio"
+      src="/src/assets/ImagenesOptimizadas/ImagenPrincipalInicio.png"
+    />
+  );
+};
+
+export default ImagenPricipaLogo;
