+++ PodoFrontGonnet/src/components/css/footer.css
